feat: Enhance incident management forms with new fields and improved data handling
- Updated CAPAFormReadOnly and EditCAPAWizard to ensure estimated cost is consistently handled as a number.
- Added incident commander and response team fields to CreateAiIncidentWizard and EditAiIncidentWizard for better incident tracking.
- Enhanced BasicInformationStep and ImpactAssessmentStep with clearer placeholder texts for user guidance.
- Renamed LinksEvidenceStep to Links & References and improved labeling for clarity.
- Updated ResponseGovernanceStep to reflect changes in classification and compliance terminology, enhancing user understanding.
