+++ components/app/incidents/create/CreateAiIncidentWizard.tsx
@@ -79,11 +79,11 @@ const CreateAiIncidentWizard: React.FC = () => {
           "domain",
           "severity",
           "status",
+          "incident_commander",
+          "response_team",
         ]);
       case 2:
         return await trigger([
-          "incident_commander",
-          "response_team",
           "primary_regulatory_framework",
           "notification_requirement",
         ]);
@@ -111,9 +111,9 @@ const CreateAiIncidentWizard: React.FC = () => {
   const getCurrentStepFields = (step: number): (keyof AiIncidentFormData)[] => {
     switch (step) {
       case 1:
-        return ["title", "summary", "incident_type", "domain", "severity", "status"];
+        return ["title", "summary", "incident_type", "domain", "severity", "status", "incident_commander", "response_team"];
       case 2:
-        return ["incident_commander", "response_team", "primary_regulatory_framework", "notification_requirement"];
+        return ["primary_regulatory_framework", "notification_requirement"];
       case 3:
         return ["estimated_impacted_records", "data_types_impacted"];
       case 4:
